refactor: 패키지 구조 변경 Layered Architecture + DDD (#12)
