+++ src/main/java/com/danjitalk/danjitalk/api/user/member/MemberController.java
@@ -1,0 +1,7 @@
+package com.danjitalk.danjitalk.api.user.member;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MemberController {
+}
