--- src/main/java/com/danjitalk/danjitalk/common/entity/BaseEntity.java
@@ -1,24 +1,0 @@
-package com.danjitalk.danjitalk.common.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public class BaseEntity {
-
-    @CreatedDate
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
