--- src/main/java/com/danjitalk/danjitalk/community/reaction/enums/ReactionType.java
@@ -1,13 +1,0 @@
-package com.danjitalk.danjitalk.community.reaction.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ReactionType {
-
-    LIKE("좋아요");
-
-    private final String description;
-}
