+++ src/main/java/com/danjitalk/danjitalk/domain/apartment/enums/ApartmentType.java
@@ -1,0 +1,16 @@
+package com.danjitalk.danjitalk.domain.apartment.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ApartmentType {
+
+    APT("아파트"),
+    VILLA("빌라"),
+    OFFICE("오피스텔");
+
+    private final String description;
+
+}
