+++ src/main/java/com/danjitalk/danjitalk/domain/chat/entity/ChatroomMemberMapping.java
@@ -1,0 +1,23 @@
+package com.danjitalk.danjitalk.domain.chat.entity;
+
+import com.danjitalk.danjitalk.domain.user.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ChatroomMemberMapping {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Chatroom chatroom;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Member member;
+
+}
