+++ src/main/java/com/danjitalk/danjitalk/domain/community/reaction/entity/Reaction.java
@@ -1,0 +1,25 @@
+package com.danjitalk.danjitalk.domain.community.reaction.entity;
+
+import com.danjitalk.danjitalk.domain.community.feed.entity.Feed;
+import com.danjitalk.danjitalk.domain.community.reaction.enums.ReactionType;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Reaction {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    private ReactionType reactionType;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "feed_id")
+    private Feed feed;
+
+}
