+++ src/main/java/com/danjitalk/danjitalk/domain/community/reaction/enums/ReactionType.java
@@ -1,0 +1,13 @@
+package com.danjitalk.danjitalk.domain.community.reaction.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ReactionType {
+
+    LIKE("좋아요");
+
+    private final String description;
+}
