+++ src/main/java/com/danjitalk/danjitalk/domain/user/member/entity/SystemUser.java
@@ -1,0 +1,33 @@
+package com.danjitalk.danjitalk.domain.user.member.entity;
+
+import com.danjitalk.danjitalk.domain.common.entity.BaseEntity;
+import com.danjitalk.danjitalk.domain.user.member.enums.LoginMethod;
+import com.danjitalk.danjitalk.domain.user.member.enums.Role;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class SystemUser extends BaseEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long systemUserId;
+
+    private String loginId;
+
+    private String password;
+
+    private LocalDateTime lastLoginTime;
+
+    @Enumerated(EnumType.STRING)
+    private LoginMethod lastLoginMethod;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+}
