--- src/main/java/com/danjitalk/danjitalk/user/user/enums/Gender.java
@@ -1,16 +1,0 @@
-package com.danjitalk.danjitalk.user.user.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@RequiredArgsConstructor
-@ToString
-public enum Gender {
-
-    MALE("남성"),
-    FEMALE("여성");
-
-    private final String description;
-}
