--- src/main/java/com/danjitalk/danjitalk/user/user/enums/Role.java
@@ -1,16 +1,0 @@
-package com.danjitalk.danjitalk.user.user.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@RequiredArgsConstructor
-@ToString
-public enum Role {
-
-    USER("사용자"),
-    ADMIN("어드민");
-
-    private final String description;
-}
